requirements.txt
streamlit>=1.32.0
plotly>=5.22.0
matplotlib>=3.8.0
pandas>=2.0.0
PyYAML>=6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.14.0
ollama>=0.1.23
langchain>=0.2.0
ragas>=0.1.11
PyPDF2>=3.0.0
python-docx>=0.8.11
tiktoken>=0.5.2
scikit-learn>=1.3.0